streaming-data-to-analytics/func_test/requirements.txt
google-auth>=2.22.0
google-cloud-monitoring>=2.22
requests>=2.31.0
pycryptodome>=3.19.0
crypto>=1.4
uuid>=1.3
Faker==13.15.1
pyjwt>=2.10
google-cloud